requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-Babel==4.0.0
reportlab==4.0.7
flask-cors==4.0.0
google-genai==0.2.2
python-dotenv==1.0.0
twilio==9.3.0

